Lock @types/node version to ^8.x
Previous syntax to specify this package's version unintentionally installs previous version in terms of semantic versioning.
This is because definition of `latest version` in NPM is the latest version in **chronological** order.
For example, if a package releases version 1.1.0 after releasing 2.0.0, package-name: '*' installs 1.1.0,
which @types/node often does.
This commit makes sure we will keep depending on 8.x.x version of @types/node.