Bump `entities` to `4.0.0`, update code
`entities` is now an ES Module, which is something that should make ES Module fans happy.

The new version features a smaller decode tree, which required an update to the tokenizer. It also includes `escapeText` and `escapeAttribute` functions that implement the required behavior for the HTML serializer; I replaced parse5's `escapeString` function with these.