Bump parse5 to 7.1.0
This version is already on npm — forgot that I should PR the release first.

Once this is merged, I'll create a tag and publish release notes.

My gut feeling is that, going forward we should skip versions for the other packages, so that versions numbers across packages are monotonically increasing.