fix: account for undefined in element stack (#1436)
* fix: handle exhaustive stack popping

In some weird cases it seems you can end up trying to pop the element
stack until you basically reach `-1` (i.e. `indexOf` failed to find what
we want).

If that happens, the current element will be `undefined`. This change
adds some handling around that.

* fix: account for `undefined` in element stack

Changes the element stack such that the current element and ID can be
`undefined`, as this is already true but not represented in the types.

Also then updates any usages to account for this.